requirements.txt
# Deep Learning & Math
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0

# Evaluation & Interpretability
scikit-learn>=1.0.0
captum>=0.6.0      # <-- NEW: Axiomatic Attribution

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0